problem-solver/py/requirements.txt
py-sc-kpm==0.4.0
py-sc-client==0.4.0
py-sc-kpm==0.4.0
py-sc-client==0.4.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
requests~=2.32.3
